Add tests for Itemlist
