+++ src/functional_components/DriveContextMenu/package.json
@@ -1,0 +1,1 @@
+{"main": "DriveContextMenu.js"}
